api code cleaning
